refactor(pages): move pages to src/pages/
* update import paths
* add default exports
* move layout to a dedicated folder
