+++ client/vite.config.ts
@@ -9,7 +9,8 @@ export default defineConfig({
   plugins: [react(), tailwindcss()],
   resolve: {
     alias: {
-      "@": path.resolve(__dirname, "./src")
+      "@": path.resolve(__dirname, "./src"),
+      "@server": path.resolve(__dirname, "../server")
     }
   },
   server: {
